Extract styles for OneTimeDonation component
